glxにRename
